Converted to Climacell.
Converted to Climacell API using requests. Also put in a Fahrenheit flag as I'm stuck in the US with terrible non SI units.
Icon lookup is prone to breaking unless we add a lot more icons for each of the weatherCode options.
Also included some fonts which weren't committed to the original codebase but were referenced.
Added an option to rotate the display.